* Ajout de la partie client * Correction d'un bug lors de l'édition d'une interevetion qui a un temps passé non rond (exemple 8:31 affiche un résultat à décimal)
git-svn-id: http://open.iabsis.com/subversion/igestis-commercial/trunk@1969 9ce8b743-0909-46e4-8230-39c73aa7ff53

+++ classes/Common/StringManipulation.php
@@ -16,7 +16,7 @@ class StringManipulation {
     public static function convertDecimalToTimeFormat($nbMinutes) {
         if($nbMinutes == 0) return "00:00";
         else {
-            $hours = $nbMinutes / 60;
+            $hours = (int)($nbMinutes / 60);
             $minutes = $nbMinutes % 60;
             while (strlen($hours) < 2) $hours = "0" . $hours;
             while (strlen($minutes) < 2) $minutes = "0" . $minutes;
